chaindeck-api-main/pyproject.toml
[tool.poetry]
name = "prometheus-fast-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "prometheus_fast_api"}]

[tool.poetry.dependencies]
python = "^3.10"
prometheus-client = "^0.15.0"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
prometheus-api-client = "^0.5.2"
requests = "^2.28.1"
supertokens-python = "^0.11.8"
nest-asyncio = "^1.5.6"
sqlalchemy = "^2.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.5"
sqlalchemy-utils = "^0.40.0"
pydantic = {extras = ["email"], version = "^1.10.6"}
clickhouse-driver = "^0.2.5"
clickhouse-connect = "^0.5.16"
web3 = "^6.0.0"
psycopg2 = "^2.9.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "runserver:main"